changed an int to size_t in dataman.h to avoid warnings

+++ src/DataMan.h
@@ -87,7 +87,7 @@ class DataMan{
                 get_callback = cb;
             }
             else{
-                for (int i=0; i<m_next.size(); i++){
+                for (size_t i=0; i<m_next.size(); i++){
                     m_next[i]->reg_callback(cb);
                 }
             }
